% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_null_pval.R
\name{calcNullPval}
\alias{calcNullPval}
\title{Calculate pval for each gene in synthetic null data}
\usage{
calcNullPval(synthetic_null, normalize = T, hvg = NULL, seed = 123, nCores = 1)
}
\arguments{
\item{synthetic_null}{A matrix (genes x cells), or a list of such matrices representing synthetic null data. If not a list, it will be coerced to a list.}

\item{normalize}{Logical; if \code{TRUE} (default), apply Seurat's NormalizeData to each dataset. If \code{FALSE}, sets assay data directly. Set to false if null data is generated using PCA approximation. It is recommended to provide hvg if normalize is false.}

\item{hvg}{A list of feaures; will be applied if normalization is set to false.}

\item{seed}{Numeric; random seed for clustering reproducibility (default: 123).}

\item{nCores}{An integer. The number of cores to use for parallel processing.}
}
\value{
A named list with preprocess data and gene p-values
}
\description{
Runs a Seurat-based null differential expression analysis for provided synthetic null datasets.
}
