% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_de.R
\name{callDE}
\alias{callDE}
\title{Call differentially expressed (DE) genes by Clipper algorithm}
\usage{
callDE(targetScores, nullScores, nlogTrans = TRUE, FDR = 0.05, correct = FALSE)
}
\arguments{
\item{targetScores}{A named numeric vector of the DE scores from the target data, e.g., the p-values between two clusters from the real data.}

\item{nullScores}{A named numeric vector of the DE scores from the synthetic null data, e.g., the p-values between two clusters from the null data.}

\item{nlogTrans}{A logical value. If the input scores are p-values, take the \code{-log10} transformation since Clipper require larger scores represent more significant DE. Default is TRUE.}

\item{FDR}{A numeric value of the target False Discovery Rate (FDR). Default is 0.05.}

\item{correct}{A logical value. If TRUE, perform the correction to make the distribution of contrast scores approximately symmetric. Default is FALSE.}
}
\description{
\code{callDE} takes two vectors representing each gene's significance. They are usually p-values from the target data and synthetic null data.
}
\details{
This function constructs the contrast scores by taking the difference between target DE scores and null DE scores.
}
\examples{
targetScores <- runif(10000)
nullScores <- runif(10000)
names(targetScores) <- names(nullScores) <- paste0("Gene", 1:10000)
res <- callDE(targetScores, nullScores, correct = FALSE)

}
