% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleCounts}
\alias{exampleCounts}
\title{A count matrix of 100 genes from the A549 cell line from Tian et al., Nature Methods, 2019.}
\format{
A dataset with 100 rows (genes) and 1256 cols (cells)
}
\usage{
data("exampleCounts")
}
\value{
The corresponding count matrix object
}
\description{
A count matrix of 100 genes from the A549 cell line from Tian et al., Nature Methods, 2019.
}
\keyword{datasets}
