% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers.R
\name{findMarkers}
\alias{findMarkers}
\title{Find differential markers using ClusterDE}
\usage{
findMarkers(obj, ident.1, ident.2, nCores = 1, nRep = 1, flavour = "classic")
}
\arguments{
\item{obj}{Seurat object containing expression data.}

\item{ident.1}{Group identity to test as group 1.}

\item{ident.2}{Group identity to test as group 2.}

\item{nCores}{Number of CPU cores used for parallelization.}

\item{nRep}{Number of null replicates.}

\item{flavour}{Either \code{"classic"} or \code{"pca"}; selects null-construction method. \strong{classic}: uses default count matrix as simulation input. \strong{pca}: uses PCA approximation as simulation input, faster and more scalable when large number of replicates is needed.}
}
\value{
A table of DEG results.
}
\description{
Find differential markers using ClusterDE
}
\examples{
dummy_obj <- DUMMY
findMarkers(dummy_obj, ident.1 = "cell_type_1", ident.2 = "cell_type_2")
}
