% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{logcp10k}
\alias{logcp10k}
\title{Log-transformed Counts per 10,000 (logCP10K)}
\usage{
logcp10k(x, scale.factor = 10000)
}
\arguments{
\item{x}{A matrix or \code{Matrix::dgCMatrix} of raw counts with genes as rows and cells as columns.}

\item{scale.factor}{A numeric value representing the scaling factor for total counts per cell (default: 1e4).}
}
\value{
A matrix of the same dimensions as \code{x}, containing log-transformed normalized counts per cell.
}
\description{
Computes the per-cell log-transformed counts per 10,000 for a matrix of raw counts.
}
