% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_de.R
\name{callDE}
\alias{callDE}
\title{Call differentially expressed (DE) genes by Clipper algorithm}
\usage{
callDE(
  targetScores,
  nullScores,
  nlogTrans = TRUE,
  FDR = 0.05,
  contrastScore = "diff",
  correct = FALSE,
  threshold = "BC",
  ordering = TRUE
)
}
\arguments{
\item{targetScores}{A named numeric vector of the DE scores from the target data, e.g., the p-values between two clusters from the real data.}

\item{nullScores}{A named numeric vector of the DE scores from the synthetic null data, e.g., the p-values between two clusters from the null data.}

\item{nlogTrans}{A logical value. If the input scores are p-values, take the \code{-log10} transformation since Clipper require larger scores represent more significant DE. Default is TRUE.}

\item{FDR}{A numeric value of the target False Discovery Rate (FDR). Must be 'diff' or 'max'.}

\item{contrastScore}{A string value of the way to construct contrast scores. The choice can be}

\item{correct}{A logical value. If TRUE, perform the correction to make the distribution of contrast scores approximately symmetric. Default is FALSE.}

\item{threshold}{A string value of the threshold method. Must be 'BC' or 'DS'.}

\item{ordering}{A logic value. If TRUE, order the genes in the returned table by their significance. Default is TRUE.}
}
\value{
A list of target FDR, DE genes, and the detailed summary table.
}
\description{
\code{callDE} takes two vectors representing each gene's significance. They are usually p-values from the target data and synthetic null data.
}
\details{
This function constructs the contrast scores by taking the difference between target DE scores and null DE scores.
}
\examples{
targetScores <- runif(10000)
nullScores <- runif(10000)
names(targetScores) <- names(nullScores) <- paste0("Gene", 1:10000)
res <- callDE(targetScores, nullScores, correct = FALSE)

}
