% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_null.R
\name{constructNull}
\alias{constructNull}
\title{Construct the synthetic null data}
\usage{
constructNull(
  mat,
  family = "nb",
  formula = NULL,
  extraInfo = NULL,
  nCores = 1,
  parallelization = "pbmcmapply",
  fastVersion = FALSE,
  corrCut = 0.1,
  BPPARAM = NULL
)
}
\arguments{
\item{mat}{An expression matrix (gene by cell). It can be a regular matrix or a \code{sparseMatrix}.}

\item{family}{A string or a vector of strings of the distribution of your data.
Must be one of 'nb', 'binomial', 'poisson', 'zip', 'zinb' or 'gaussian', which represent 'poisson distribution',
'negative binomial distribution', 'zero-inflated poisson distribution', 'zero-inflated negative binomail distribution',
and 'gaussian distribution' respectively. For UMI-counts data, we usually use 'nb'. Default is 'nb'.}

\item{formula}{A string of the mu parameter formula. It defines the relationship between gene expression in synthetic null data and the extra covariates. Default is NULL.
For example, if your input data is a spatial data with X, Y coordinates, the formula can be 's(X, Y, bs = 'gp', k = 4)'.}

\item{extraInfo}{A data frame of the extra covariates used in \code{formula}. Default is NULL.}

\item{nCores}{An integer. The number of cores to use for Parallel processing.}

\item{parallelization}{A string indicating the specific parallelization function to use.
Must be one of 'mcmapply', 'bpmapply', or 'pbmcmapply', which corresponds to the parallelization function in the package
\code{parallel},\code{BiocParallel}, and \code{pbmcapply} respectively. The default value is 'pbmcmapply'.}

\item{fastVersion}{A logic value. If TRUE, the fast approximation is used. Default is FALSE.}

\item{corrCut}{A numeric value. The cutoff for non-zero proportions in genes used in modelling correlation.}

\item{BPPARAM}{A \code{MulticoreParam} object or NULL. When the parameter parallelization = 'mcmapply' or 'pbmcmapply',
this parameter must be NULL. When the parameter parallelization = 'bpmapply',  this parameter must be one of the
\code{MulticoreParam} object offered by the package 'BiocParallel. The default value is NULL.}
}
\value{
The expression matrix of the synthetic null data.
}
\description{
\code{constructNull} takes the target data as the input and returns the corresponding synthetic null data.
}
\details{
This function constructs the synthetic null data based on the target data (real data). The input is a expression matrix (gene by cell); the user should specify a distribution, which is usually Negative Binomial for count matrix.
}
\examples{
data(exampleCounts)
nullData <- constructNull(mat = exampleCounts)

}
